Abilities API: Normalize input from schema
Without this patch REST API would require a weird empty `?input` field for optional input given how the current controller works with input schema when it defines the expected shape. This patch normalizes the input for the ability, applying the default value from the input schema when needed.

Developed in https://github.com/WordPress/wordpress-develop/pull/10395.

Follow-up [61032], [61045].

Props gziolo, jorgefilipecosta, mukesh27.
Fixes #64139.



Built from https://develop.svn.wordpress.org/trunk@61047


git-svn-id: http://core.svn.wordpress.org/trunk@60383 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/rest-api/endpoints/class-wp-rest-abilities-v1-categories-controller.php
@@ -93,7 +93,7 @@ class WP_REST_Abilities_V1_Categories_Controller extends WP_REST_Controller {
 		$offset   = ( $page - 1 ) * $per_page;
 
 		$total_categories = count( $categories );
-		$max_pages        = ceil( $total_categories / $per_page );
+		$max_pages        = (int) ceil( $total_categories / $per_page );
 
 		if ( $request->get_method() === 'HEAD' ) {
 			$response = new WP_REST_Response( array() );
